+++ fixtures/time-zone.ts
@@ -44,7 +44,8 @@ export const cbSaveSanitizedSettingsEditMessageTextPayload = (): unknown => ({
     ],
   },
   text:
-    "<b>Time Zone</b>\nI can work in different time zones and display dates in the appropriate format.\n\n" +
+    "<b>Time Zone</b>\n" +
+    "I can work in different time zones and display dates in the appropriate format.\n\n- - -\n\n" +
     `Select a time zone for @${supergroup.username ?? ""} chat.\n\nCurrent time zone: <b>GMT+0 Etc/UTC</b>\n` +
     `Modified at ${formatInTimeZone(Date.now(), "UTC", DATE_FORMAT)} ` +
     `by <a href="tg:user?id=${adminUser.id.toString()}">@${adminUser.username ?? ""}</a>`,
@@ -74,7 +75,7 @@ export const cbSaveSettingsEditMessageTextPayload = (): unknown => ({
   },
   text:
     "<b>Time Zone</b>\n" +
-    "I can work in different time zones and display dates in the appropriate format.\n\n" +
+    "I can work in different time zones and display dates in the appropriate format.\n\n- - -\n\n" +
     `Select a time zone for @${supergroup.username ?? ""} chat.\n\n` +
     `Current time zone: <b>${formatInTimeZone(new Date(), "Europe/London", "zzz")} Europe/London</b>\n` +
     `Modified at ${formatInTimeZone(Date.now(), "UTC", DATE_FORMAT)} ` +
@@ -107,6 +108,60 @@ export const cbSaveSettingsWebhook = {
     message: { chat: privateChat, date: Date.now(), edit_date: Date.now(), from: bot, message_id: 1, text: "" },
   },
   update_id: 1,
+};
+
+/**
+ * Payload for edit message text request. It should be sent as a result of settings callback.
+ */
+export const cbSettingsEditMessageTextPayload = {
+  chat_id: privateChat.id,
+  message_id: 1,
+  parse_mode: "HTML",
+  reply_markup: {
+    inline_keyboard: [
+      expect.arrayContaining([]),
+      expect.arrayContaining([]),
+      expect.arrayContaining([]),
+      expect.arrayContaining([]),
+      expect.arrayContaining([]),
+      [{ callback_data: `${TimeZoneAction.SETTINGS}?cId=${supergroup.id.toString()}&s=5`, text: "»" }],
+      [{ callback_data: `${SettingsAction.FEATURES}?cId=${supergroup.id.toString()}`, text: backToFeaturesText }],
+    ],
+  },
+  text:
+    "<b>Time Zone</b>\nI can work in different time zones and display dates in the appropriate format.\n\n- - -\n\n" +
+    `Select a time zone for @${supergroup.username ?? ""} chat.\n\nCurrent time zone: <b>GMT+0 UTC</b>`,
+};
+
+/**
+ * Payload for edit message text request. It should be sent as a result of settings callback on a specific page.
+ */
+export const cbSettingsPageEditMessageTextPayload = {
+  chat_id: privateChat.id,
+  message_id: 1,
+  parse_mode: "HTML",
+  reply_markup: {
+    // Necessary for expectations
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
+    inline_keyboard: expect.arrayContaining([
+      [
+        expect.objectContaining({
+          callback_data: `${TimeZoneAction.SAVE}?cId=${supergroup.id.toString()}&v=Europe%2FLondon`,
+        }),
+      ],
+      [expect.objectContaining({ text: "«" }), expect.objectContaining({ text: "»" })],
+      [
+        {
+          callback_data: `${SettingsAction.FEATURES}?cId=${supergroup.id.toString()}`,
+          text: backToFeaturesText,
+        },
+      ],
+    ]),
+  },
+  text:
+    "<b>Time Zone</b>\nI can work in different time zones and display dates in the appropriate format.\n\n- - -\n\n" +
+    `Select a time zone for @${supergroup.username ?? ""} chat.\n\nCurrent time zone: ` +
+    `<b>${formatInTimeZone(new Date(), "Europe/London", "zzz")} Europe/London</b>`,
 };
 
 /**